Move carousel related methods to extension
