+++ Sources/iOS/Classes/Spot.swift
@@ -1,7 +1,6 @@
 // swiftlint:disable weak_delegate
 
 import UIKit
-import Tailor
 
 public class Spot: NSObject, Spotable, SpotHorizontallyScrollable {
 
@@ -149,56 +148,6 @@ public class Spot: NSObject, Spotable, SpotHorizontallyScrollable {
     }
   }
 
-  fileprivate func setupHorizontalCollectionView(_ collectionView: CollectionView, with size: CGSize) {
-    guard let layout = collectionView.collectionViewLayout as? GridableLayout else {
-      return
-    }
-
-    collectionView.isScrollEnabled = true
-    layout.scrollDirection = .horizontal
-    #if os(iOS)
-      collectionView.isPagingEnabled = component.interaction.paginate == .page
-    #endif
-    configurePageControl()
-
-    if collectionView.contentSize.height > 0 {
-      collectionView.frame.size.height = collectionView.contentSize.height
-    } else {
-      var newCollectionViewHeight: CGFloat = 0.0
-
-      newCollectionViewHeight <- component.items.sorted(by: {
-        $0.size.height > $1.size.height
-      }).first?.size.height
-
-      collectionView.frame.size.height = newCollectionViewHeight
-
-      if collectionView.frame.size.height > 0 {
-        collectionView.frame.size.height += layout.sectionInset.top + layout.sectionInset.bottom
-      }
-    }
-
-    configureCollectionViewHeader(collectionView, with: size)
-
-    CarouselSpot.configure?(collectionView, layout)
-
-    collectionView.frame.size.height += layout.headerReferenceSize.height
-
-    if let componentLayout = component.layout {
-      collectionView.frame.size.height += CGFloat(componentLayout.inset.top + componentLayout.inset.bottom)
-    }
-
-    if let pageIndicatorPlacement = component.layout?.pageIndicatorPlacement {
-      switch pageIndicatorPlacement {
-      case .below:
-        layout.sectionInset.bottom += pageControl.frame.height
-        pageControl.frame.origin.y = collectionView.frame.height
-      case .overlay:
-        let verticalAdjustment = CGFloat(2)
-        pageControl.frame.origin.y = collectionView.frame.height - pageControl.frame.height - verticalAdjustment
-      }
-    }
-  }
-
   fileprivate func layoutCollectionView(_ collectionView: CollectionView, with size: CGSize) {
     prepareItems()
 
@@ -210,17 +159,6 @@ public class Spot: NSObject, Spotable, SpotHorizontallyScrollable {
     }
   }
 
-  fileprivate func layoutHorizontalCollectionView(_ collectionView: CollectionView, with size: CGSize) {
-    guard let collectionViewLayout = collectionView.collectionViewLayout as? GridableLayout else {
-      return
-    }
-
-    collectionViewLayout.prepare()
-    collectionViewLayout.invalidateLayout()
-    collectionView.frame.size.width = size.width
-    collectionView.frame.size.height = collectionViewLayout.contentSize.height
-  }
-
   func registerDefaultIfNeeded(view: View.Type) {
     guard Configuration.views.storage[Configuration.views.defaultIdentifier] == nil else {
       return
@@ -229,7 +167,7 @@ public class Spot: NSObject, Spotable, SpotHorizontallyScrollable {
     Configuration.views.defaultItem = Registry.Item.classType(view)
   }
 
-  private func configurePageControl() {
+  func configurePageControl() {
     guard let placement = component.layout?.pageIndicatorPlacement else {
       pageControl.removeFromSuperview()
       return
